move files to top directory
