+++ heidgaf_log_collector/server.py
@@ -5,7 +5,7 @@ import queue
 import sys  # needed for Terminal execution
 
 sys.path.append(os.getcwd())  # needed for Terminal execution
-from pipeline_prototype.heidgaf_log_collector import utils
+from heidgaf_log_collector import utils
 from pipeline_prototype.logging_config import setup_logging
 
 MAX_NUMBER_OF_CONNECTIONS = 5
